chore: fix i18n issue and test
