+++ web/src/components/molecules/Settings/FormModal.tsx
@@ -5,6 +5,7 @@ import Form, { Rule } from "@reearth-cms/components/atoms/Form";
 import Input from "@reearth-cms/components/atoms/Input";
 import Modal from "@reearth-cms/components/atoms/Modal";
 import Select from "@reearth-cms/components/atoms/Select";
+import useSettings from "@reearth-cms/components/molecules/Settings/useSettings.ts";
 import {
   TileType,
   TerrainType,
@@ -12,7 +13,7 @@ import {
   TerrainInput,
   WorkspaceSettings,
 } from "@reearth-cms/components/molecules/Workspace/types";
-import { useT, t } from "@reearth-cms/i18n";
+import { useT } from "@reearth-cms/i18n";
 import { newID } from "@reearth-cms/utils/id";
 import { validateURL } from "@reearth-cms/utils/regex";
 
@@ -23,21 +24,6 @@ type FormValues = {
   image?: string;
   cesiumIonAssetId?: string;
   cesiumIonAccessToken?: string;
-};
-
-export const TileTypeFormat: Record<TileType, string> = {
-  DEFAULT: "Google Satellite",
-  ROAD_MAP: "Google Road Map",
-  OPEN_STREET_MAP: "OpenStreetMap",
-  EARTH_AT_NIGHT: "NASA Black Marble",
-  JAPAN_GSI_STANDARD_MAP: t("Japan GSI Standard Map"),
-  URL: "URL",
-};
-
-export const TerrainTypeFormat: Record<TerrainType, string> = {
-  // CESIUM_WORLD_TERRAIN: "Cesium World Terrain",
-  REEARTH_TERRAIN: "Re:Earth Terrain",
-  CESIUM_ION: "Cesium Ion",
 };
 
 type Props = {
@@ -62,6 +48,7 @@ const FormModal: React.FC<Props> = ({
   const t = useT();
   const [form] = Form.useForm<FormValues>();
   const [extraOpen, setExtraOpen] = useState(false);
+  const { TileTypeFormat, TerrainTypeFormat } = useSettings();
 
   const options = useMemo(() => {
     const typeFormat = isTile ? TileTypeFormat : TerrainTypeFormat;
@@ -69,7 +56,7 @@ const FormModal: React.FC<Props> = ({
       value: key,
       label: typeFormat[key as keyof typeof typeFormat],
     }));
-  }, [isTile]);
+  }, [TerrainTypeFormat, TileTypeFormat, isTile]);
 
   useEffect(() => {
     if (open) {
